+++ Assets/Scripts/EnemyMov.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 19f9263d57a56af429bf54e8ad3e36b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
