+++ Assets/Scripts/Gun.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3b173a9930afd564fbaa37e813b20655
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
