+++ Assets/Scripts/PlayerMov.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerMov : MonoBehaviour
+{
+    public float speed = 5f;
+
+    private void Update()
+    {
+        ApplyMovement();
+    }
+
+    void ApplyMovement()
+    {
+        var horizontal = Input.GetAxis("Horizontal");
+        var vertical = Input.GetAxis("Vertical");
+
+        transform.Translate(Time.deltaTime * speed * new Vector3(horizontal, vertical));
+    }
+}
