+++ Assets/Scripts/PlayerMov.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1209facec2dbae6419a9745899dc8572
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
