+++ Assets/Scripts/Spaceship.cs
@@ -4,7 +4,6 @@ public class Spaceship : MonoBehaviour
 {
     private SpriteRenderer _renderer;
     private Bounds _cameraBounds;
-    public Projectile projectilePrefab;
     public float speed = 5f;
 
     void Start()
@@ -28,24 +27,7 @@ public class Spaceship : MonoBehaviour
 
     void Update()
     {
-        ApplyMovement();
-        FireProjectile();
-    }
-
-    
-
-    void FireProjectile()
-    {
-        if (Input.GetButtonDown("Fire1")) {
-            Instantiate(projectilePrefab, transform.position, Quaternion.identity);
-        }
-    }
-
-    void ApplyMovement()
-    {
-        var horizontal = Input.GetAxis("Horizontal");
-        var vertical = Input.GetAxis("Vertical");
-
-        transform.Translate(Time.deltaTime * speed * new Vector3(horizontal, vertical));
+        //ApplyMovement();
+        //FireProjectile();
     }
 }
